Fix white space errors
 - Fix white space errors
 - Fix linting errors
 - Move classes to their own files
