+++ dankbot/meme.py
@@ -1,0 +1,7 @@
+class Meme(object):
+    """
+    Model for meme objects
+    """
+    def __init__(self, link, source):
+        self.link = link
+        self.source = source
